fix the size of the templates as I forgot to resize them before committing previously,  added a few road markings and grass features
In the previous commit I forgot to resize the size of templates so that each only took up as only as much space as they needed to.
added a few items to the grass items theme, these still need to be turned into isometric and made to look better.
added some road marking to the road markings theme, still have to add these to the templates.